% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_03.R
\name{euclidean}
\alias{euclidean}
\title{function for euclid or euclideans algorithm
to find out the GCD of two numbers}
\usage{
euclidean(num1, num2)
}
\arguments{
\item{num1}{is a number or an integer}

\item{num2}{is a number or an integer
both parameters are inputs for euclidean
to find out their GCD}
}
\value{
the GCD(greatest common divisor)
}
\description{
function for euclid or euclideans algorithm
to find out the GCD of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
