% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_data}
\alias{wiki_data}
\title{Dataset creation for wiki graph.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{edge1 of graph}
  \item{v2}{edge2 of grpah}
  \item{w}{weight of the edge}
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph}
}
\usage{
wiki_data
}
\description{
A dataset containing the edges of the graph from (v1 to v2) with the weight
of edge (w)
}
\keyword{datasets}
